# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dengue",
        "description": "Documentação da API em Laravel com Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/denuncias": {
            "get": {
                "tags": [
                    "Denuncia"
                ],
                "summary": "Lista todas as denúncias",
                "operationId": "f73bb1b07b8e31f2bb9a6796caaddd70",
                "responses": {
                    "200": {
                        "description": "Lista de denúncias retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Denuncia"
                ],
                "summary": "Cadastra uma nova denúncia",
                "operationId": "24673d8163545379da36426235bd14af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "IdUsuario",
                                    "TipoFoco",
                                    "Descricao",
                                    "Longitude",
                                    "Latitude",
                                    "Status"
                                ],
                                "properties": {
                                    "IdUsuario": {
                                        "type": "integer"
                                    },
                                    "TipoFoco": {
                                        "type": "string"
                                    },
                                    "Descricao": {
                                        "type": "string"
                                    },
                                    "Longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "Latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "Status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Denúncia cadastrada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao cadastrar denúncia"
                    }
                }
            }
        },
        "/api/denuncias/{id}": {
            "get": {
                "tags": [
                    "Denuncia"
                ],
                "summary": "Exibe uma denúncia específica",
                "operationId": "e5e49bea73989a289b7d9f238b7264cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da denúncia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Denúncia retornada com sucesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Denuncia"
                ],
                "summary": "Atualiza uma denúncia existente",
                "operationId": "d1da1ce4dd5699e6c2b0f1d90e972d27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da denúncia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "IdUsuario",
                                    "Descricao",
                                    "Endereco",
                                    "Status"
                                ],
                                "properties": {
                                    "IdUsuario": {
                                        "type": "integer"
                                    },
                                    "Descricao": {
                                        "type": "string"
                                    },
                                    "Endereco": {
                                        "type": "string"
                                    },
                                    "Status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Denúncia atualizada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao atualizar denúncia"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Denuncia"
                ],
                "summary": "Remove uma denúncia",
                "operationId": "0177bbf6b8e2f3d0cab652bcf5f841c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da denúncia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Denúncia removida com sucesso"
                    },
                    "400": {
                        "description": "Erro ao remover denúncia"
                    }
                }
            }
        },
...
    },
...
}